add StarRating component to SharedComponent folder
