+++ client/src/components/rr/ReviewTile.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 //import exampleDataList from './exampleDataList.js';
-//import StarRating from './StarRating.jsx';
+import StarRating from '../SharedComponent/StarRating.jsx';
 
 const Tile = styled.div`
   border: 1px solid #ccc;
@@ -10,67 +10,67 @@ const Tile = styled.div`
 `;
 
 
-const StarContainer = styled.div`
-  display: inline-block;
-  font-size: 24px;
-`;
+// const StarContainer = styled.div`
+//   display: inline-block;
+//   font-size: 24px;
+// `;
 
-const Star = styled.span`
-  color: ${({ isActive }) => (isActive ? 'gold' : 'gray')};
-`;
+// const Star = styled.span`
+//   color: ${({ isActive }) => (isActive ? 'gold' : 'gray')};
+// `;
 
-const FullStar = styled(Star)`
-  color: gold;
-`;
+// const FullStar = styled(Star)`
+//   color: gold;
+// `;
 
-const HalfStar = styled(Star)`
-  position: relative;
-  color: gray;
+// const HalfStar = styled(Star)`
+//   position: relative;
+//   color: gray;
 
-  &::before {
-    content: '\2605';
-    position: absolute;
-    width: 50%;
-    overflow: hidden;
-    color: gold;
-    z-index: 1;
-    left: 0;
-    top: 0;
-  }
-`;
-
-
-const StarRating = ({ rating, totalStars }) => {
-  const fullStars = Math.floor(rating);
-  const decimalPart = rating - fullStars;
-
-   //console.log('rating is: ', rating);
-  //  console.log(fullStar);
-  //  console.log(decimalPart);
-
-  const renderStar = (index) => {
-
-    console.log('index:', index, 'fullStars:', fullStars, 'decimalPart:', decimalPart);
-
-    if (index < fullStars) {
-      return <FullStar key={index} isActive>&#9733;</FullStar>;
-    } else if (index === fullStars) {
-      if (decimalPart >= 0.75) {
-        return <FullStar key={index} isActive>&#9733;</FullStar>;
-      } else if (decimalPart >= 0.25) {
-        return <HalfStar key={index} isActive>&#9733;</HalfStar>;
-      } else {
-        return <Star key={index}>&#9734;</Star>;
-      }
-    } else {
-      return <Star key={index}>&#9734;</Star>;
-    }
+//   &::before {
+//     content: '\2605';
+//     position: absolute;
+//     width: 50%;
+//     overflow: hidden;
+//     color: gold;
+//     z-index: 1;
+//     left: 0;
+//     top: 0;
+//   }
+// `;
 
 
-  };
+// const StarRating = ({ rating, totalStars }) => {
+//   const fullStars = Math.floor(rating);
+//   const decimalPart = rating - fullStars;
 
-  return <StarContainer>{[...Array(totalStars)].map((star, index) => renderStar(index))}</StarContainer>;
-};
+//    //console.log('rating is: ', rating);
+//   //  console.log(fullStar);
+//   //  console.log(decimalPart);
+
+//   const renderStar = (index) => {
+
+//     console.log('index:', index, 'fullStars:', fullStars, 'decimalPart:', decimalPart);
+
+//     if (index < fullStars) {
+//       return <FullStar key={index} isActive>&#9733;</FullStar>;
+//     } else if (index === fullStars) {
+//       if (decimalPart >= 0.75) {
+//         return <FullStar key={index} isActive>&#9733;</FullStar>;
+//       } else if (decimalPart >= 0.25) {
+//         return <HalfStar key={index} isActive>&#9733;</HalfStar>;
+//       } else {
+//         return <Star key={index}>&#9734;</Star>;
+//       }
+//     } else {
+//       return <Star key={index}>&#9734;</Star>;
+//     }
+
+
+//   };
+
+//   return <StarContainer>{[...Array(totalStars)].map((star, index) => renderStar(index))}</StarContainer>;
+// };
 
 const Summary = styled.div`
   font-weight: bold;
@@ -118,7 +118,7 @@ const ReviewTile = ({ review }) => {
 
   return (
     <Tile>
-      <StarRating rating={rating} totalStars={5} />
+      <StarRating rating={rating} />
       <Summary>{summary}</Summary>
       <Recommend>{recommend ? 'I recommend this product' : ''}</Recommend>
       <Body>{body}</Body>
